add the rest of the extensions
