+++ src/system/ExtensionsModule/Twig/Extension/DefaultPathExtension.php
@@ -13,48 +13,16 @@ declare(strict_types=1);
 
 namespace Zikula\ExtensionsModule\Twig\Extension;
 
-use Symfony\Component\Routing\RouterInterface;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
-use Zikula\ExtensionsModule\Api\ApiInterface\CapabilityApiInterface;
+use Zikula\ExtensionsModule\Twig\Runtime\DefaultPathRuntime;
 
 class DefaultPathExtension extends AbstractExtension
 {
-    /**
-     * @var CapabilityApiInterface
-     */
-    private $capabilityApi;
-
-    /**
-     * @var RouterInterface
-     */
-    private $router;
-
-    public function __construct(
-        CapabilityApiInterface $capabilityApi,
-        RouterInterface $router
-    ) {
-        $this->capabilityApi = $capabilityApi;
-        $this->router = $router;
-    }
-
     public function getFunctions()
     {
         return [
-            new TwigFunction('getDefaultPath', [$this, 'getDefaultPath']),
+            new TwigFunction('getDefaultPath', [DefaultPathRuntime::class, 'getDefaultPath']),
         ];
-    }
-
-    public function getDefaultPath(string $extensionName, $type = CapabilityApiInterface::USER): string
-    {
-        $capability = $this->capabilityApi->isCapable($extensionName, $type);
-        if (!$capability) {
-            return '';
-        }
-        if (isset($capability['route'])) {
-            return $this->router->generate($capability['route']);
-        }
-
-        return '';
     }
 }
