+++ src/system/ExtensionsModule/Twig/Extension/UserInterfaceExtension.php
@@ -13,31 +13,18 @@ declare(strict_types=1);
 
 namespace Zikula\ExtensionsModule\Twig\Extension;
 
-use Symfony\Component\HttpKernel\Fragment\FragmentHandler;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
-use Zikula\ExtensionsModule\Twig\Extension\SimpleFunction\ModuleFooterFunction;
-use Zikula\ExtensionsModule\Twig\Extension\SimpleFunction\ModuleHeaderFunction;
-use Zikula\ExtensionsModule\Twig\Extension\SimpleFunction\ModuleLinksFunction;
+use Zikula\ExtensionsModule\Twig\Runtime\UserInterfaceRuntime;
 
 class UserInterfaceExtension extends AbstractExtension
 {
-    /**
-     * @var FragmentHandler
-     */
-    private $handler;
-
-    public function __construct(FragmentHandler $handler)
-    {
-        $this->handler = $handler;
-    }
-
     public function getFunctions()
     {
         return [
-            new TwigFunction('moduleHeader', [new ModuleHeaderFunction($this->handler), 'display'], ['is_safe' => ['html']]),
-            new TwigFunction('moduleLinks', [new ModuleLinksFunction($this->handler), 'display'], ['is_safe' => ['html']]),
-            new TwigFunction('moduleFooter', [new ModuleFooterFunction($this->handler), 'display'], ['is_safe' => ['html']])
+            new TwigFunction('moduleHeader', [UserInterfaceRuntime::class, 'moduleHeader'], ['is_safe' => ['html']]),
+            new TwigFunction('moduleLinks', [UserInterfaceRuntime::class, 'moduleLinks'], ['is_safe' => ['html']]),
+            new TwigFunction('moduleFooter', [UserInterfaceRuntime::class, 'moduleFooter'], ['is_safe' => ['html']])
         ];
     }
 }
