+++ src/system/ThemeModule/Twig/Extension/BrandingExtension.php
@@ -13,103 +13,19 @@ declare(strict_types=1);
 
 namespace Zikula\ThemeModule\Twig\Extension;
 
-use Exception;
-use Twig\Environment;
 use Twig\Extension\AbstractExtension;
 use Twig\TwigFunction;
-use Zikula\Bundle\CoreBundle\Site\SiteDefinitionInterface;
-use Zikula\ThemeModule\Engine\Asset;
+use Zikula\ThemeModule\Twig\Runtime\BrandingRuntime;
 
 class BrandingExtension extends AbstractExtension
 {
-    /**
-     * @var Environment
-     */
-    private $twig;
-
-    /**
-     * @var SiteDefinitionInterface
-     */
-    private $site;
-
-    /**
-     * @var Asset
-     */
-    private $assetHelper;
-
-    public function __construct(
-        Environment $twig,
-        SiteDefinitionInterface $site,
-        Asset $assetHelper
-    ) {
-        $this->twig = $twig;
-        $this->site = $site;
-        $this->assetHelper = $assetHelper;
-    }
-
     public function getFunctions()
     {
         return [
-            new TwigFunction('siteName', [$this, 'getSiteName']),
-            new TwigFunction('siteSlogan', [$this, 'getSiteSlogan']),
-            new TwigFunction('siteBranding', [$this, 'getSiteBrandingMarkup'], ['is_safe' => ['html']]),
-            new TwigFunction('siteImagePath', [$this, 'getSiteImagePath'])
+            new TwigFunction('siteName', [BrandingRuntime::class, 'getSiteName']),
+            new TwigFunction('siteSlogan', [BrandingRuntime::class, 'getSiteSlogan']),
+            new TwigFunction('siteBranding', [BrandingRuntime::class, 'getSiteBrandingMarkup'], ['is_safe' => ['html']]),
+            new TwigFunction('siteImagePath', [BrandingRuntime::class, 'getSiteImagePath'])
         ];
-    }
-
-    /**
-     * Returns site name.
-     */
-    public function getSiteName(): string
-    {
-        return $this->site->getName();
-    }
-
-    /**
-     * Returns site slogan.
-     */
-    public function getSiteSlogan(): string
-    {
-        return $this->site->getSlogan();
-    }
-
-    /**
-     * Returns site branding markup.
-     */
-    public function getSiteBrandingMarkup(): string
-    {
-        return $this->twig->render('@ZikulaThemeModule/Engine/manifest.html.twig');
-    }
-
-    /**
-     * Returns site image path.
-     */
-    public function getSiteImagePath(string $imageType = ''): string
-    {
-        if (!in_array($imageType, ['logo', 'mobileLogo', 'icon'], true)) {
-            $imageType = 'logo';
-        }
-
-        $accessor = 'get' . ucfirst($imageType) . 'Path';
-
-        $assetPath = $this->site->{$accessor}();
-
-        try {
-            $imagePath = $this->assetHelper->resolve($assetPath);
-        } catch (Exception $exception) {
-            // fall back to default
-            $assetPath = '@CoreBundle:images/';
-            if ('logo' === $imageType) {
-                $assetPath .= 'logo_with_title.png';
-            } elseif ('mobileLogo' === $imageType) {
-                $assetPath .= 'zk-power.png';
-            } elseif ('icon' === $imageType) {
-                $assetPath .= 'logo.gif';
-            }
-
-            $imagePath = $this->assetHelper->resolve($assetPath);
-        }
-
-        return $imagePath;
     }
 }
